Check RunAsDaemon before using wait()
Avoid hangs by checking for RunAsDaemon 1 in the torrc before using
wait() on a launched tor process. Iff it's absent, use poll().

A configurable poll() time is provided in env['poll_launch_time'],
and a default in env['poll_launch_time_default'] is used when the torrc
is unexpectedly missing RunAsDaemon 1.
(TODO: We may only need one of these?)