Force authorities to vote Exit & Guard for all nodes
Use the new tor TestingDirAuthVoteExit option, and the
TestingDirAuthVoteGuard option, to make all authorities vote all relays
(and authorities) as exits and guards. This ensures that tor's
make test-network will work. But it's a sledgehammer approach.

Requires either the new tor TestingDirAuthVoteExit option, or the
issue-13161-check-torrc-options patch for chutney.