Add bridge clients and hidden services to chutney verify
Chutney verify now verifies bridge clients and hidden services.

Bridge clients are verified just like clients, by using an exit to
connect to the TrafficTester's listening port.

Hidden services are verified by using an arbitrary client
(the first one in the list) to connect to the hidden service.
The hidden service is configured to redirect data to the
TrafficTester's listening host:port.

This works almost exactly like the standard client / bridge client tests.
The only difference is that the client connects via a hidden service port
pre-configured with a specific redirect, rather than via an exit which
could handle arbitrary hosts and ports.

Also review the number of relays/exits/authorities required for the 25%
guard changes. For 2 guards, 8 relays are required in total,
including authorities and exits.