If $PYTHON isn't set, use python2 or python
If $PYTHON isn't set, use "command" to check for python2, then python.
(python3 check is disabled due to #16904 - chutney doesn't work with
python3.)

If neither of the "command" checks work, assume a strange shell or "command",
and fall back to python2.

Fixes default install of chutney on OS X, which has neither $PYTHON set,
nor python2; but does have python, which happens to be version 2.

Resolves #16902.