Fix up chutney verifier's notion of a test suite
Previously we counted the number of successes and failures, and
compared them to the number of expected tests.  That's fragile, if
we forget to count a test or if we count a success more than once.
Now we give every test a name, and we only allow tests to succeed or
fail once.