TorNet: Check each node's dir info distribution
Check that every relay has published its directory info to the
authorities, it is in the consensus, and every node has downloaded it.

Check that every bridge has published its directory info to the
bridge authority, and it is in the bridge networkstatus.

Check that every bridge client has downloaded bridge descriptors.

This code currently does not work for these chutney networks:
* bridges-min
* bridges+ipv6-min

All other networks pass the new tests.

Part of 33379.