Templating: Make _DictWrapper wrap a dict
Motivation behind this: I needed to use an `if 'key' in object`
construct and was unable to do so because _DictWrapper, despite its
name, does not inherit members from the builtin dict class. So, make it
more true to its name, and while we're at it we don't need a private
_dict member; we can just make use of dict's default functions as
applicable.