TorNet: simplify print_bootstrap_status
In order to simplify code while preserving atomicity of bootstrap status
between calls of getBootstrapStatus, introduce a new function,
updateBootstrapStatus, which stores last status in the
LocalNodeController object. That way, subsequent calls to
getBootstrapStatus are guaranteed to be same until next update, and we
have fewer lists to create and zip, fewer arguments to pass to
print_bootstrap_status, and hopefully easier code to reason about.